+++ lib/src/api/helper/rpc_hellper.dart
@@ -3,6 +3,7 @@ import 'package:polkadot_dart/src/api/core/api.dart';
 import 'package:polkadot_dart/src/provider/methods/state/call.dart';
 import 'package:polkadot_dart/src/provider/provider/provider.dart';
 
+/// Extension to assist with runtime calls based on metadata
 extension RuntimeCallHelper on MetadataApi {
   Future<T> runtimeCall<T>(
       {required String apiName,
@@ -16,8 +17,8 @@ extension RuntimeCallHelper on MetadataApi {
         methodName: methodName,
         params: params,
         fromTemplate: fromTemplate);
-    final rpcMethod = SubstrateRPCStateCall(
-        method: method, data: BytesUtils.toHexString(encode, prefix: "0x"));
+    final encodeInputs = BytesUtils.toHexString(encode, prefix: "0x");
+    final rpcMethod = SubstrateRPCStateCall(method: method, data: encodeInputs);
     final response = await rpc.request(rpcMethod);
     return decodeRuntimeApiOutput(
         apiName: apiName,