+++ lib/src/models/generic/generic.dart
@@ -9,3 +9,4 @@ export 'models/sp_weights_weight_v2_weight.dart';
 export 'models/account_data.dart';
 export 'models/runtime_version.dart';
 export 'models/runtime_version_api.dart';
+export 'models/frame_support_dispatch_per_dispatch_class.dart';