+++ lib/src/models/generic/models/sp_weights_weight_v2_weight.dart
@@ -2,12 +2,16 @@ import 'package:blockchain_utils/layout/core/core/core.dart';
 import 'package:polkadot_dart/polkadot_dart.dart';
 import 'package:polkadot_dart/src/models/generic/layouts/layouts.dart';
 
+/// sp_weights, weight_v2, Weigh
 class SpWeightsWeightV2Weight
     extends SubstrateSerialization<Map<String, dynamic>> {
   final BigInt refTime;
   final BigInt proofSize;
   const SpWeightsWeightV2Weight(
       {required this.refTime, required this.proofSize});
+  SpWeightsWeightV2Weight.deserializeJson(Map<String, dynamic> json)
+      : refTime = json["ref_time"],
+        proofSize = json["proof_size"];
 
   @override
   Layout<Map<String, dynamic>> layout({String? property}) {